+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Fix crashitem names mismatch between client and server.
   (`#172 <https://github.com/pytest-dev/pytest-rerunfailures/issues/172>`_)
 
+- Fix crash when setup fails with --rerun-except flag.
+  (`#230 <https://github.com/pytest-dev/pytest-rerunfailures/issues/230>`_)
 
 12.0 (2023-07-05)
 -----------------